+++ src/public/mathfield.ts
@@ -82,13 +82,15 @@ export type InsertOptions = {
 export interface Mathfield {
     mode: ParseMode;
 
-    getConfig<K extends keyof MathfieldConfig>(
+    // @newapi getOptions, getOption
+    getConfig?<K extends keyof MathfieldConfig>(
         keys: K[]
     ): Pick<MathfieldConfig, K>;
-    getConfig<K extends keyof MathfieldConfig>(key: K): MathfieldConfig[K];
-    getConfig(): MathfieldConfig;
+    getConfig?<K extends keyof MathfieldConfig>(key: K): MathfieldConfig[K];
+    getConfig?(): MathfieldConfig;
 
-    $setConfig(config: Partial<MathfieldConfig>): void;
+    // @newapi setOptions
+    $setConfig?(config: Partial<MathfieldConfig>): void;
 
     /**
      * Reverts this mathfield to its original content.
@@ -99,7 +101,7 @@ export interface Mathfield {
      * To turn the element back into a mathfield, call
      * `MathLive.makeMathField()` on the element again to get a new mathfield object.
      */
-    $revertToOriginalContent(): void;
+    $revertToOriginalContent?(): void;
 
     /**
      * Performs a command defined by a selector.
@@ -119,7 +121,8 @@ export interface Mathfield {
      * In the above example, both calls invoke the same selector.
      *
      */
-    $perform(command: Selector | [Selector, ...any[]]): boolean;
+    // @newapi executeCommand
+    $perform?(command: Selector | [Selector, ...any[]]): boolean;
 
     /**
      * Returns a textual representation of the mathfield.
@@ -129,7 +132,8 @@ export interface Mathfield {
      * @return {string}
      * @category Accessing the Content
      */
-    $text(format?: OutputFormat): string;
+    // @newapi getValue(format)
+    $text?(format?: OutputFormat): string;
 
     /**
      * Returns a textual representation of the selection in the mathfield.
@@ -138,11 +142,12 @@ export interface Mathfield {
      * **Default** = `"latex"`
      * @category Accessing the Content
      */
-    $selectedText(format?: OutputFormat): string;
+    $selectedText?(format?: OutputFormat): string;
 
-    $select(): void;
+    // @newapi select()
+    $select?(): void;
 
-    $clearSelection(): void;
+    $clearSelection?(): void;
 
     /**
      * Checks if the selection is collapsed.
@@ -150,7 +155,7 @@ export interface Mathfield {
      * @return True if the length of the selection is 0, that is, if it is a single
      * insertion point.
      */
-    $selectionIsCollapsed(): boolean;
+    $selectionIsCollapsed?(): boolean;
 
     /**
      * Returns the depth of the selection group.
@@ -161,17 +166,17 @@ export interface Mathfield {
      * which is at the root level, return 1. Note that in that case, the numerator
      * would be the "selection group".
      */
-    $selectionDepth(): number;
+    $selectionDepth?(): number;
 
     /**
      * Checks if the selection starts at the beginning of the selection group.
      */
-    $selectionAtStart(): boolean;
+    $selectionAtStart?(): boolean;
 
     /**
      * Checks if the selection extends to the end of the selection group.
      */
-    $selectionAtEnd(): boolean;
+    $selectionAtEnd?(): boolean;
     /**
      * Sets or gets the content of the mathfield.
      *
@@ -183,14 +188,16 @@ export interface Mathfield {
      *
      * @category Accessing the Content
      */
-    $latex(text?: string, options?: InsertOptions): string;
+    // @newapi getValue(options)
+    // @newapi setValue(value, options)
+    $latex?(text?: string, options?: InsertOptions): string;
 
     /**
      * Return the DOM element associated with this mathfield.
      *
      * Note that `this.$el().mathfield === this`
      */
-    $el(): HTMLElement;
+    $el?(): HTMLElement;
 
     /**
      * Inserts a block of text at the current insertion point.
@@ -204,20 +211,21 @@ export interface Mathfield {
      *
      * @category Changing the Content
      */
-    $insert(s: string, options?: InsertOptions): boolean;
+    // @newapi insert(options)
+    $insert?(s: string, options?: InsertOptions): boolean;
 
     /**
      * @category Focus
      */
-    $hasFocus(): boolean;
+    $hasFocus?(): boolean;
     /**
      * @category Focus
      */
-    $focus(): void;
+    $focus?(): void;
     /**
      * @category Focus
      */
-    $blur(): void;
+    $blur?(): void;
 
     /**
      * Updates the style (color, bold, italic, etc...) of the selection or sets
@@ -232,7 +240,8 @@ export interface Mathfield {
      * If there is no selection, the style will apply to the next character typed.
      *
      */
-    $applyStyle(style: Style): void;
+    // @newapi applyStyle(options)
+    $applyStyle?(style: Style): void;
 
     /**
      * @param keys - A string representation of a key combination.
@@ -249,7 +258,7 @@ export interface Mathfield {
      * @return Return true if the field need to be re-rendered
      * @category Changing the Content
      */
-    $keystroke(keys: string, evt?: KeyboardEvent): boolean;
+    $keystroke?(keys: string, evt?: KeyboardEvent): boolean;
 
     /**
      * Simulates a user typing the keys indicated by text.
@@ -257,7 +266,7 @@ export interface Mathfield {
      * @param text - A sequence of one or more characters.
      * @category Changing the Content
      */
-    $typedText(text: string): void;
+    $typedText?(text: string): void;
 
     getCaretPosition(): { x: number; y: number } | null;
     setCaretPosition(x: number, y: number): boolean;