+++ src/public/mathlive.ts
@@ -19,12 +19,14 @@
  */
 
 import { Mathfield } from './mathfield';
+import { MathfieldElement } from './mathfield-element';
 import { MathfieldConfig, TextToSpeechOptions } from './config';
 import { MacroDictionary, ErrorListener, ParserErrorCode } from './core';
 import { ErrorCode as MathJsonErrorCode } from '../math-json/public';
 
 export { Mathfield };
 export { MathfieldConfig };
+export { MathfieldElement };
 
 /**
  * Current version: `{{SDK_VERSION}}`